+++ sample-expo/android/app/build.gradle
@@ -87,9 +87,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace 'com.smileidentity.react.expo.example'
+    namespace 'com.smileidentity.react.expo.sample'
     defaultConfig {
-        applicationId 'com.smileidentity.react.expo.example'
+        applicationId 'com.smileidentity.react.expo.sample'
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
