+++ sample-expo/android/app/src/main/java/com/smileidentity/react/expo/sample/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.smileidentity.react.expo.example
+package com.smileidentity.react.expo.sample
 
 import android.app.Application
 import android.content.res.Configuration
